+++ core/ChaosActuatorInterface.h
@@ -49,8 +49,8 @@ namespace chaos {
             } ChaosActuatorOpcode;
             
             typedef struct {
-                double fvalue0;
-                double fvalue1;
+                float fvalue0;
+                float fvalue1;
 		char str[MAX_STR_SIZE];
 
                 int32_t ivalue;
@@ -60,8 +60,8 @@ namespace chaos {
             } actuator_iparams_t;
             
             typedef struct {
-                double fvalue0;
-                double fvalue1;
+                float fvalue0;
+                float fvalue1;
                 int32_t ivalue;
                 int32_t result;
                 uint64_t alarm_mask;
@@ -102,7 +102,7 @@ namespace chaos {
                  @param readingType the position reading method to be used to
                  @return 0 if success or an error code
                  */
-                int getPosition(readingTypes readingType,double *deltaPosition_mm);
+                int getPosition(readingTypes readingType,float *deltaPosition_mm);
                 
                 int resetAlarms(uint64_t alrm);
