+++ core/CmdACTMoveRelative.cpp
@@ -56,7 +56,7 @@ void own::CmdACTMoveRelative::setHandler(c_data::CDataWrapper *data) {
 	int err = 0;
 	int state;
         int *tmpInt;
-        double position;
+        float position;
 	std::string state_str;
 	float offset_mm = 0.f;
 	chaos::common::data::RangeValueInfo attr_info;
@@ -188,7 +188,7 @@ void own::CmdACTMoveRelative::setHandler(c_data::CDataWrapper *data) {
 void own::CmdACTMoveRelative::acquireHandler() {
 	int err = 0;
         int state=0;
-        double position;
+        float position;
         
 	std::string state_str;
 	//acquire the current readout
